Add cell coords in cli printed grid + fix cell drawing in case vals have more than 1 digit
Whenever the grid is printed to the CLI, also the row and col coordinates are printed, so that it is easier for the user to refer to them.
Also, the space inside each cell is computed as a function of the number of digits of the largest value that can be printed in the cell, so that the grid can be drawn for any value of the block/grid size.